requirements.txt
langchain
langchain-community
langchain_experimental
langchain-core
langchain-groq
langchain-huggingface
langchain-pinecone
pinecone[grpc]
pypdf
python-dotenv
sentence-transformers
flask
gunicorn
ipykernel
# -e .